Renamed package to mpl_scatter_density
