+++ mpl_scatter_density/color.py
