+++ mpl_scatter_density/scatter_density_artist.py
