+++ mpl_scatter_density/tests/baseline/1.5.x/test_downres.png
