+++ mpl_scatter_density/tests/baseline/2.0.x/test_remove.png
