+++ mpl_scatter_density/tests/test_scatter_density_artist.py
