+++ setup.py
@@ -1,16 +1,13 @@
-import os
-
-import numpy as np
-from setuptools import setup
-from setuptools.extension import Extension
+from setuptools import setup, find_packages
 
 
-setup(name='raster-axes',
+setup(name='mpl-scatter-density',
       version='0.1.dev0',
+      description='Matplotlib helpers to make density scatter plots',
+      long_description=open('README.rst').read(),
       install_requires=['numpy', 'matplotlib', 'fast-histogram'],
       author='Thomas Robitaille',
       author_email='thomas.robitaille@gmail.com',
       license='BSD',
-      url='https://github.com/astrofrog/scatter-density',
-      packages=['raster_axes']
-)
+      url='https://github.com/astrofrog/mpl-scatter-density',
+      packages=find_packages())
